refac: useState 파일 분리
